requirements.txt
flask
gunicorn
traceback2
click
itsdangerous
Jinja2
MarkupSafe
itsdangerous
Jinja2
Werkzeug
json
flask_pymongo
requests
smtplib
dnspython